AddConfig using reflection
This adds a new method to the configuration utils, that allows you to
dynamically register config objects from public members of an arbitrary type.

It also modifies the old AddConfig method to use this new one, return
the config object (to make configuration easier), and accept subtypes of
VersionedConfig instead of BaseConfig.