Binary buffer (#55)
Methods to serialize and deserialize datapoints and events to a simple binary format. Using a fixed format like this makes it very compact and fast. Deserialization uses a stream for efficiency. Points are stored as [id][point][point]..., so ids may be duplicated in a file. This is not a very stable format, and it may not work well if moved between systems.

Strings are encoded as UTF-8, and are prefixed by their length.