Error handling (#65)
Prototype for error handling in assets, we need to make sure that this is indeed how we want to handle this.

This should really be tested with some proper integration tests against CDF.

This is intended as doing the full business of correctly handling errors from CDF and returning a useful, readable result. There is a little overhead in checking for duplicates before pushing, but the error handling doesn't cause any overhead unless something goes wrong.

We don't attempt to "fix" data, just remove the offending items. If a request fails with no clear offending item, we assume that the error is fatal and return.

Each extension method takes a "RetryMode" option, which indicates how retries should be done. Supported modes are "None", which only returns the first error, if any, "OnError", which removes bad objects as described above and retries, "OnFatal", which does the same as OnError, but retries automatically on 500s or similar, and "OnError/OnFatal-KeepDuplicates", which in GetOrCreate waits until duplicates are found in CDF.