Integration tests (#71)
Add framework for integration tests, and add a few for assets, timeseries and events.

The most important features to test are handling of CDF errors, since this behavior may change unexpectedly. The old mock-tests are kept for now, since it is difficult to replicate some behavior, and it doesn't hurt to keep them.