Dispose LiteDatabase object
This generally should not affect implementations, since we use a
singleton, and it is disposed with the service provider.

Notably, this caused tests to fail on Windows, which uses a stricter
file-lock than unix-based systems.

This is unlikely to affect usage as the intended pattern only ever keeps
a single copy. Tests are affected, however, as the file lock is not
removed between tests.