Improve the task throttler (#91)
An investigation of the task throttler revealed some flaws. This aims to improve those, make the solution a bit more rigorous, and close some potential holes.

This simplifies the task throttler a bit, since it did some very not useful things, which added unecessary complexity. This is more than enough for what it is likely to be used for.