Avoid using BlockingCollection to signal termination (#99)
Turns out the previous PR didn't really fix the issue, only create a new one. The main issue seems to have been using BlockingCollection.CompleteAdding to signal that the run loop should terminate. It wouldn't always, which in long running applications that create many TaskSchedulers (like tests), would mean that the thread pool filled up with internal waiting in the blocking collection.

Over time this causes C# to run out of threads, which causes it to allocate new ones (the max is something like 1000), which is slow. This adds heavy load which in my case caused some bad timeouts and extremely inconsistent behavior.

This approach is completely different and should be much safer. Instead of always both waiting for task completion and new available tasks, we now:

 - Wait for available tasks if scheduling new ones is allowed.
 - Else wait for tasks to terminate with timeout if _timeunit is set (since in those cases allowed tasks change over time)
 - Else wait for tasks to terminate

All these are interrupted only by a global cancellationtokensource which is canceled once the throttler is closed, which should guarantee that no tasks are left behind.

Hopefully this is a bit sturdier. Testing so far seems to indicate that it is.