Base extractor class (#109)
Base class for extractors, to minimize boilerplate code when creating simple extractors.

This is probably just the first step, for live data. Ideally it should also have features designed for history, and some way to enable state-store without too much boilerplate.

This PR also adds the `PeriodicScheduler` class, which simply runs tasks with a scheduled period, which can be interrupted externally. This is something more advanced extractors often end up needing, and it creates a very nice decentralized system for reading form source systems. In this case it is used heavily in the base extractor.

There's also a new main example to the README, which is also in a separate (working) example project.