Fix periodic scheduler
This was a tricky issue. Something was wrong with the way we did
async/await, mixing synchronous/async code and Func. Not sure exactly
why this caused problems, but the solution is to clearly distinguish
between passed tasks running on the thread pool and synchronously, not
allowing the passed task to be either. This is considered the "proper"
way to do this.