Refactor CogniteError to take a generic parameter for "Skipped"
Also remove "Data", it can be added to the generic type in "Skipped" if
needed. "Skipped" will be renamed. The issue is that it is far too hard
for a user to know what the contents of an error actually is, without
using a debugger or something else. This isn't ideal, and results in bad
code.