Task scheduler (#107)
A fairly abstract construct that I nonetheless ended up needing several times in the OPC-UA extractor and decided to add here, as part of an effort to reduce code duplication in OPC-UA. If it doesn't fit I can just add it to the extractor instead, but I think it is sufficiently general that it doesn't hurt to keep it here.

The idea is that it's a sort of producer-consumer scheduler, where a generic list of items are consumed. Each item may need to be run several times, and may produce a new list of child items to consume.

An example of this is exploring a hierarchy, but also reading history (in that case no new nodes are discovered, but this is still a useful construct).

In addition, it contains a system to share capacity. In OPC-UA this is a system of continuation points, but it could also be a cross-source throttler if history involves reading from multiple sources.

This is intended as a very generic base class, it might be helpful to add other base classes later, to handle more specific cases.