Installer template (#110)
The idea is to make it as simple as possible to create a .NET installer. It isn't simple, so that involves a decent bit of creative configuration. The end result is a JSON configuration file which must be filled in with some seemingly unnecessary information that I couldn't figure out how to extract from the various project files. These are fed automatically into the .wixproj file.

The build script takes "version" and "description", which are embedded into the executable and expected to be unique to each build (In the OPC-UA extractor, which this is based on, version is a tag, and description is the git commit hash + time), location of the MSBuild executable, and path to the config file.

I've tried to make modifying the template simple as well, but dynamic and conditional XML is not very forgiving, so it is a bit of a mess.

The extractor can be added as a windows service, with optional parameters. This does not work that well with how this is done for PI, where there is a separate service and standalone version, but it works well for OPC-UA where these are merged, and once we add systems for doing the same to the utils it will work well for simple extractors as well.