Add option to automatically gzip datapoints (#118)
The default value and common in config.example.yml is based on some research done while writing this for the SDK. Turns out you need a fairly large number of datapoints before it is worthwhile. This feature counts the number of datapoints (not really the number of bytes, we could count lengths of all strings but I'm uncertain of how much that could cost), and decides on whether or not to use gzip based on a configurable limit.