Improve error handling in base extractor class (#115)
It produced some ridiculous errors even when just terminating. The problem with async is that it produces just AggregateException, which always needs to be caught and searched if you're looking for a specific exception.

This PR also adds a callback to do config validation/manual modification on startup. This is very often necessary.