Refactor errors (#117)
The original use of CogniteError and CogniteResult was just for assets/timeseries/events, and there it worked well enough. Now with more types, the use of `object` makes the system difficult to use. This PR changes that so that CogniteError now takes a type parameter, and also standardizing what an error represents.

This PR also fixes a few issues with logging, and adds a "ThrowOnError" method to CogniteResult so that errors can be translated back into exceptions.