Scheduler error handling (#125)
Improved ability to handle errors in PeriodicScheduler. It was too difficult to know whether a task existed or not, and what to do about it. This adds the option to pass tasks without name, and check whether a task with a given name exists.

It also fixes a bug where thrown errors were not properly reported.