Add system to use cron expressions in scheduler (#175)
Adds a new CronTimeSpanWrapper which is compatible with the old TimeSpanWrapper, but also supports cron expressions.

Also makes it so that you can pass a timespan provider to PeriodicScheduler instead of a timespan, making it possible to pass a TimeSpanWrapper or CronTimeSpanWrapper directly, so you can use it to schedule cron jobs as well.

Running tasks with cron timing is something that is often useful for extractors, and has been requested for OPC-UA. This makes migrating from existing TimeSpanWrapper based implementations trivial and backwards compatible. (Replace TimeSpanWrapper with CronTimeSpanWrapper, and pass the wrapper itself instead of .Value to the scheduler).