Schema (#196)
Author schema definitions for extractor utils.

We should look into a way to host these publicly, so that we can retrieve them as part of CICD.

To make writing these bearable, we must be able to split them into separate files. When storing them in the database we probably would like to download and merge them. A solution could be to use the `$defs` keyword:

 - Scan the document for `$ref`s
 - Load all referenced documents, either from the internet or from local files.
 - Store each referenced item in `$defs`.
 - Replace remote refs with local refs in generated document.

This is doable, and relatively clean.

The schema can be referenced quite elegantly from github, so long as we tag releases:

`https://raw.githubusercontent.com/cognitedata/dotnet-extractor-utils/v1.4.0/README.md` is a public, plaintext reference to the readme file at version 1.4.0.