Peek value instead of doing explicit null check
This is a very different approach, but the previous behavior was
arguably caused by a bug. We consumed the value from the YAML document
if it was a `string`, independent of whether or not we did string
substitution on it.

This meant that we consumed nulls as well, instead of passing them along
to the null literal serializer.

It might have had other unexpected side effects as well.