Base extractor (#522)
First take on a new base extractor class. Some more work goes into this in the next step, adding startup reporting and integrating things a bit more tightly, but this is a nice first draft that contains some of the core logic.

The base extractor has a few roles, currently:

 - It contains the config object, in the future it will also handle changes to config revision and restart.
 - It acts as an error reporter, for raw errors.
 - It manages "monitored tasks", which are essentially just C# Tasks. The idea is that if something fails in a background process, this needs to be properly reported to the extractor so that it can take appropriate action. Background tasks have a name, can be cancelled independently, and can be expected to end or not, so if a task ends without an error, the extractor can take appropriate action.
 - It also contains shutdown logic, which is highly customizable, since extractor implementations may want to shut components down in a specific order. In general, different parts of the process are supposed to be independently cancelled, so that graceful shutdown is possible.

As with the last iteration of the base extractor, it is intended to be used as a base class for user extractors.

* Comments from review

---------

Co-authored-by: Ozan Göktan <72358629+ozangoktan@users.noreply.github.com>