Improve the checkin worker a bit, add a callback for config changes
This is necessary step, making the checkin worker capable of calling
some callback to inform the runtime and/or extractor (not sure yet) that
a new config revision has been produced, and that the extractor should
restart.

It will be up to the outer loop to decide whether the config changes are
significant enough to warrant a change (if we decide to go that way).