Improve the checkin worker a bit, add a callback for config changes (#529)
This is necessary step, making the checkin worker capable of calling some callback to inform the runtime and/or extractor (not sure yet) that a new config revision has been produced, and that the extractor should restart.

It will be up to the outer loop to decide whether the config changes are significant enough to warrant a change (if we decide to go that way).

I also improved the checkin worker a bit in general, making it a bit more resilient to cancellation.

It's quite important that errors aren't lost when the extractor shuts down, so we need to make sure that any errors that failed to write or weren't written are kept for a future checkin.