Report startup (#531)
Final part of startup reporting. This ensures that we actually send the startup request before we report any checkins, thus avoiding inconsistent state in CDF, even if the extractor runs offline for some time before being able to report startup.

The ability to run offline is opt-in, so the checkin worker can be configured to crash if we cannot connect to CDF.