Various things for final runtime implementation
The runtime is too big to fit in under 500 lines, this adds some
disjointed things that are necessary for it. I couldn't find a way to
split the runtime method itself, which is way too complex, so I split of
everything else.

That does mean parts of this PR makes little sense, since there is no
context to be had.