Add auth using middleware instead of through the SDK
This is a test, initially, though we may want to do it this way in the
SDK as well. By adding auth through middleware instead of on top of the
client, we should be able to automatically retry 401s, which could avoid
hard errors when the token expires while retrying.