Various things for final runtime implementation (#538)
The runtime is too big to fit in under 500 lines, this adds some disjointed things that are necessary for it. I couldn't find a way to split the runtime method itself, which is way too complex, so I split off everything else.

That does mean parts of this PR makes little sense, since there is no context to be had.

This makes a few fixes to the builder, and adds two sink implementations. One is for the bootstrap process, where before we load the config and start periodic checkin, we still need to be able to report errors. Reuse the IIntegrationSink type for this for now, it may make sense to change that a bit later, perhaps split the interface into two.

The other is for the case where the extractor is running without a connection to CDF at all, which is possible for some extractors. In this case we still need a sink, so the log sink simply logs all incoming events and does nothing else with them.