Add auth using middleware instead of through the SDK (#540)
By adding auth through middleware instead of on top of the client, we should be able to automatically retry 401s, which could avoid hard errors when the token expires while retrying.

This makes it so that:
  - We retry requests in the authenticator.
  - We get a new token if it expires before we retry a request to CDF.

Which should really help with avoiding spurious errors.

The reason why this is done here instead of in the SDK is that the SDK is fundamentally built to leave configuring the http client to the user, and this must be done before the http client is created.

Since we're in the middle of rewriting the library (in the unstable module), there are some duplicated changes in this PR.