CDF mocking framework (#563)
We have various ways to do this in different extractors, and it would be
nice to have a common framework. The plan is to extend the tests in the
utils to also use this.

There's some unpleasant DI magic to get this to work properly with the
extensions we add to HTTP clients, but it works fairly well, and is
hopefully flexible enough that it can be expanded for greater ease of
use. In theory we could even create pre-built mocks for common CDF APIs.

Co-authored-by: MUHAMMED YASEEN <121yaseen@gmail.com>