Use periodic scheduler in unstable base extractor
Required some refactoring and new features in the scheduler, but I think
the result is pretty good. There's no real reason to have _two_ of
these, and this way it's possible to use the "periodic" features on the
periodic scheduler directly from extractors, which we do a fair bit,
simplifying migration.